chore: move group test to enterprise
